+++ src/ga4gh/gks/metaschema/tools/release_prep/cli.py
@@ -18,61 +18,13 @@ from ga4gh.gks.metaschema.tools.config import (
     SUPPRESS_UNSUPPORTED_KEY_WARNING_ENV,
     load_metaschema_config,
 )
-from ga4gh.gks.metaschema.tools.release_prep.git import (
-    CommandRunner,
-    SubmoduleUpdate,
-)
-from ga4gh.gks.metaschema.tools.release_prep.git import (
-    get_product_repo_dir as _get_product_repo_dir,
-)
-from ga4gh.gks.metaschema.tools.release_prep.git import (
-    infer_submodule_update as _infer_submodule_update,
-)
-from ga4gh.gks.metaschema.tools.release_prep.git import (
-    infer_submodule_update_from_current_branch as _infer_submodule_update_from_current_branch,
-)
-from ga4gh.gks.metaschema.tools.release_prep.git import (
-    require_upstream_branch_when_submodule_exists as _require_upstream_branch_when_submodule_exists,
-)
-from ga4gh.gks.metaschema.tools.release_prep.git import (
-    update_submodule as _update_submodule,
-)
-from ga4gh.gks.metaschema.tools.release_prep.git import (
-    validate_submodule as _validate_submodule,
-)
-from ga4gh.gks.metaschema.tools.release_prep.product_config import (
-    get_schema_build_dir as _get_schema_build_dir,
-)
-from ga4gh.gks.metaschema.tools.release_prep.product_config import (
-    infer_product_from_repo_dir as _infer_product_from_repo_dir,
-)
-from ga4gh.gks.metaschema.tools.release_prep.product_config import (
-    resolve_product_dir as _resolve_product_dir,
-)
-from ga4gh.gks.metaschema.tools.release_prep.product_config import (
-    update_product_version as _update_product_version,
-)
-from ga4gh.gks.metaschema.tools.release_prep.schema_versions import (
-    main as update_schema_versions,
-)
-from ga4gh.gks.metaschema.tools.release_prep.worktree import (
-    CommandOutputRunner,
-    Reporter,
-)
-from ga4gh.gks.metaschema.tools.release_prep.worktree import (
-    require_clean_worktree as _require_clean_worktree,
-)
-from ga4gh.gks.metaschema.tools.release_prep.worktree import (
-    warn_if_product_branch_not_current as _warn_if_product_branch_not_current,
-)
-from ga4gh.gks.metaschema.tools.release_prep.worktree import (
-    warn_if_worktree_dirty as _warn_if_worktree_dirty,
-)
+from ga4gh.gks.metaschema.tools.release_prep import git, product_config, schema_versions, worktree
 
 SOURCE_UPDATE_CHECK_FLAG = "--check"
 SOURCE_UPDATE_DISALLOW_VERSIONED_REFS_FLAG = "--disallow-versioned-refs"
 SOURCE_UPDATE_COMMAND_NAME = "source2updated"
 MAKE_ALL_COMMAND = ("make", "all")
+MAKE_CLEAN_COMMAND = ("make", "clean")
 
 
 def _run_command(command: list[str], cwd: Path) -> None:
@@ -113,7 +65,7 @@ class ReleasePrepSummary:
     product: str
     version: str
     product_dir: Path
-    submodules: list[SubmoduleUpdate]
+    submodules: list[git.SubmoduleUpdate]
     validated_only: bool = False
 
 
@@ -122,9 +74,9 @@ def _start_release(
     product: str,
     version: str,
     repo_dir: Path,
-    submodules: list[SubmoduleUpdate],
-    output_runner: CommandOutputRunner,
-    reporter: Reporter | None,
+    submodules: list[git.SubmoduleUpdate],
+    output_runner: worktree.CommandOutputRunner,
+    reporter: worktree.Reporter | None,
     fail_on_dirty: bool,
 ) -> Path:
     """Validate common release inputs and return the product schema directory.
@@ -143,10 +95,10 @@ def _start_release(
     """
     _validate_submodule_count(submodules)
     _report(reporter, f"{action} release for product {product} version {version}")
-    product_dir = _resolve_product_dir(repo_dir, product)
+    product_dir = product_config.resolve_product_dir(repo_dir, product)
     _report(reporter, f"Using product schema: {product_dir}")
     _handle_dirty_worktree(
-        _get_product_repo_dir(product_dir),
+        git.get_product_repo_dir(product_dir),
         f"Product {product}",
         output_runner,
         reporter,
@@ -156,7 +108,7 @@ def _start_release(
     return product_dir
 
 
-def _report(reporter: Reporter | None, message: str) -> None:
+def _report(reporter: worktree.Reporter | None, message: str) -> None:
     """Report release-prep progress when a reporter is configured.
 
     :param reporter: Optional progress reporter, such as ``print``.
@@ -178,7 +130,7 @@ def _run_source_update(product_dir: Path, check: bool) -> int:
     if check:
         argv.insert(0, SOURCE_UPDATE_CHECK_FLAG)
 
-    return update_schema_versions(argv)
+    return schema_versions.main(argv)
 
 
 def _suppress_repeated_config_warnings() -> str | None:
@@ -218,7 +170,7 @@ def _source_update_label(check: bool) -> str:
     return " ".join([SOURCE_UPDATE_COMMAND_NAME, *flags])
 
 
-def _validate_submodule_count(submodules: list[SubmoduleUpdate]) -> None:
+def _validate_submodule_count(submodules: list[git.SubmoduleUpdate]) -> None:
     """Validate release prep only targets the immediate upstream submodule.
 
     :param submodules: Requested submodule updates.
@@ -234,9 +186,9 @@ def _validate_submodule_count(submodules: list[SubmoduleUpdate]) -> None:
 
 def _warn_if_downstream_branch_not_current(
     product_dir: Path,
-    submodules: list[SubmoduleUpdate],
-    output_runner: CommandOutputRunner,
-    reporter: Reporter | None,
+    submodules: list[git.SubmoduleUpdate],
+    output_runner: worktree.CommandOutputRunner,
+    reporter: worktree.Reporter | None,
 ) -> None:
     """Warn about product branch freshness for downstream releases.
 
@@ -252,14 +204,14 @@ def _warn_if_downstream_branch_not_current(
     if not submodules:
         return
 
-    _warn_if_product_branch_not_current(_get_product_repo_dir(product_dir), output_runner, reporter)
+    worktree.warn_if_product_branch_not_current(git.get_product_repo_dir(product_dir), output_runner, reporter)
 
 
 def _handle_dirty_worktree(
     repo_dir: Path,
     label: str,
-    output_runner: CommandOutputRunner,
-    reporter: Reporter | None,
+    output_runner: worktree.CommandOutputRunner,
+    reporter: worktree.Reporter | None,
     fail_on_dirty: bool,
 ) -> None:
     """Warn or fail when a git working tree has uncommitted changes.
@@ -273,20 +225,20 @@ def _handle_dirty_worktree(
         uncommitted changes.
     """
     if fail_on_dirty:
-        _require_clean_worktree(repo_dir, label, output_runner)
+        worktree.require_clean_worktree(repo_dir, label, output_runner)
         return
 
-    _warn_if_worktree_dirty(repo_dir, label, output_runner, reporter)
+    worktree.warn_if_worktree_dirty(repo_dir, label, output_runner, reporter)
 
 
 def validate_release(
     product: str,
     version: str,
     repo_dir: Path = Path("."),
-    submodules: list[SubmoduleUpdate] | None = None,
-    runner: CommandRunner = _run_command,
-    output_runner: CommandOutputRunner = _run_command_output,
-    reporter: Reporter | None = None,
+    submodules: list[git.SubmoduleUpdate] | None = None,
+    runner: git.CommandRunner = _run_command,
+    output_runner: worktree.CommandOutputRunner = _run_command_output,
+    reporter: worktree.Reporter | None = None,
     fail_on_dirty: bool = False,
 ) -> ReleasePrepSummary:
     """Validate release-prep inputs without mutating the working tree.
@@ -318,11 +270,11 @@ def validate_release(
         fail_on_dirty,
     )
     load_metaschema_config(product_dir / METASCHEMA_FN)
-    resolved_submodules: list[SubmoduleUpdate] = []
+    resolved_submodules: list[git.SubmoduleUpdate] = []
 
     for submodule in requested_submodules:
         _report(reporter, f"Validating submodule {submodule.identifier} on branch {submodule.branch}")
-        _submodule_dir, _entry, resolved_submodule = _validate_submodule(
+        _submodule_dir, _entry, resolved_submodule = git.validate_submodule(
             submodule,
             product_dir,
             runner,
@@ -346,10 +298,10 @@ def prepare_release(
     product: str,
     version: str,
     repo_dir: Path = Path("."),
-    submodules: list[SubmoduleUpdate] | None = None,
-    runner: CommandRunner = _run_command,
-    output_runner: CommandOutputRunner = _run_command_output,
-    reporter: Reporter | None = None,
+    submodules: list[git.SubmoduleUpdate] | None = None,
+    runner: git.CommandRunner = _run_command,
+    output_runner: worktree.CommandOutputRunner = _run_command_output,
+    reporter: worktree.Reporter | None = None,
     fail_on_dirty: bool = False,
 ) -> ReleasePrepSummary:
     """Prepare source and generated files for a product release.
@@ -357,7 +309,7 @@ def prepare_release(
     The function mutates the product ``metaschema.yaml`` by setting the local
     product version, may update the immediate upstream git submodule, updates
     source YAML version references before build validation runs, runs
-    ``make all`` to regenerate artifacts, then verifies the source YAML files
+    ``make clean`` and ``make all`` to regenerate artifacts, then verifies the source YAML files
     are release-ready.
 
     :param product: Product directory/version key.
@@ -388,12 +340,12 @@ def prepare_release(
         fail_on_dirty,
     )
     config_fp = product_dir / METASCHEMA_FN
-    resolved_submodules: list[SubmoduleUpdate] = []
+    resolved_submodules: list[git.SubmoduleUpdate] = []
 
     for submodule in requested_submodules:
         _report(reporter, f"Updating submodule {submodule.identifier} on branch {submodule.branch}")
         resolved_submodules.append(
-            _update_submodule(
+            git.update_submodule(
                 submodule,
                 product_dir,
                 runner,
@@ -405,7 +357,7 @@ def prepare_release(
         _report(reporter, f"Checked out submodule {submodule.identifier} tag {resolved_submodules[-1].tag}")
 
     _report(reporter, f"Updating {config_fp} version {product}={version}")
-    _update_product_version(config_fp, product, version)
+    product_config.update_product_version(config_fp, product, version)
 
     _report(reporter, "Updating source YAML version references")
     update_exit = _run_source_update(product_dir, check=False)
@@ -414,8 +366,11 @@ def prepare_release(
         raise ValueError(msg)
     previous_warning_suppression = _suppress_repeated_config_warnings()
     try:
-        _report(reporter, f"Running make all in {_get_schema_build_dir(product_dir)}")
-        runner(list(MAKE_ALL_COMMAND), _get_schema_build_dir(product_dir))
+        build_dir = product_config.get_schema_build_dir(product_dir)
+        _report(reporter, f"Running make clean in {build_dir}")
+        runner(list(MAKE_CLEAN_COMMAND), build_dir)
+        _report(reporter, f"Running make all in {build_dir}")
+        runner(list(MAKE_ALL_COMMAND), build_dir)
 
         _report(reporter, "Verifying source YAML version references")
         check_exit = _run_source_update(product_dir, check=True)
@@ -508,16 +463,16 @@ def main(argv: list[str] | None = None) -> int:
         raise ValueError(msg)
 
     repo_dir = Path(".").resolve()
-    product = _infer_product_from_repo_dir(repo_dir)
-    product_dir = _resolve_product_dir(repo_dir, product)
+    product = product_config.infer_product_from_repo_dir(repo_dir)
+    product_dir = product_config.resolve_product_dir(repo_dir, product)
     submodules = None
 
     if args.upstream_branch:
-        submodules = [_infer_submodule_update(product_dir, args.upstream_branch, args.upstream_tag)]
+        submodules = [git.infer_submodule_update(product_dir, args.upstream_branch, args.upstream_tag)]
     elif args.use_current_upstream_branch:
-        submodules = [_infer_submodule_update_from_current_branch(product_dir, args.upstream_tag)]
+        submodules = [git.infer_submodule_update_from_current_branch(product_dir, args.upstream_tag)]
     elif not args.skip_upstream:
-        _require_upstream_branch_when_submodule_exists(product_dir)
+        git.require_upstream_branch_when_submodule_exists(product_dir)
 
     release_fn = validate_release if args.validate else prepare_release
     summary = release_fn(
